Adds several major GPU, dataset, cli option changes
In this work we properly assign data to the GPU device during
training. We also add more command line options to checkpoint the
save the model to disk, create an epoch loss training graph,
create an alignment figure, evaluate the model's BLEU score.
We also improve logging, ensuring that details regarding the dataset
are included in the logs. We also save log files to disk. Each
time a model is trained, all relevant data is saved to a folder
including the models parameters.